Experiment for Deep Learning Binary Classifier
* create binary classifier using deep learning
* restructure packages
